[RHTAP-1321] Method separation
