+++ pkg/utils/tekton/signing_secret.go
@@ -1,0 +1,36 @@
+package tekton
+
+import (
+	"context"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func (k KubeController) CreateOrUpdateSigningSecret(publicKey []byte, name, namespace string) (err error) {
+	api := k.Tektonctrl.KubeInterface().CoreV1().Secrets(namespace)
+	ctx := context.TODO()
+
+	expectedSecret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{Name: name},
+		Data:       map[string][]byte{"cosign.pub": publicKey},
+	}
+
+	s, err := api.Get(ctx, name, metav1.GetOptions{})
+	if err != nil {
+		if !errors.IsNotFound(err) {
+			return
+		}
+		if _, err = api.Create(ctx, expectedSecret, metav1.CreateOptions{}); err != nil {
+			return
+		}
+	} else {
+		if string(s.Data["cosign.pub"]) != string(publicKey) {
+			if _, err = api.Update(ctx, expectedSecret, metav1.UpdateOptions{}); err != nil {
+				return
+			}
+		}
+	}
+	return
+}
